+++ CHANGELOG.md
@@ -33,8 +33,8 @@
 - Using Python argparse to parse commandline arguments
 
 ## [Unreleased]
-### Changed
+### Added 
 - Added option to _index_ command to specify which columns to use as data source
 - Supporting DCS Scratchpad steerpoint with multiple coordinate formats
 - Supporting DCS Scratchpad steerpoint with elevation only in feet
-
+- Added possibility to load aerodrome information using # command in scratchpad file